Unigram model updated less often to reduce noise, added a "qualified" parameter to unigram nodes to speed up qualification branches

+++ src/kernel_lion_decode.h
@@ -100,8 +100,8 @@ typedef struct {
     uint_fast8_t lastUnigram;
 
     bool deepMode;
-    uint_fast64_t deepModeBits;
-    uint_fast64_t plainModeBits;
+    uint_fast32_t deepModeBits;
+    uint_fast32_t plainModeBits;
 
     density_lion_dictionary dictionary;
 } density_lion_decode_state;
